app/core/config.py
from dotenv import load_dotenv
import os
from pathlib import Path

BASE_DIR = Path(__file__).resolve().parent.parent.parent
load_dotenv(BASE_DIR / ".env")
# load_dotenv()

DATABASE_URL = os.getenv("DATABASE_URL")